Convert test_integration to asyncio to avoid testing threading races
Fixes #768
